requirements.txt
Flask==3.0.2
google_cloud_aiplatform==1.38.1
ipython==8.18.1
protobuf==4.25.2
pytube==15.0.0
vertexai==0.0.1
youtube_transcript_api==0.6.2
gunicorn
google-cloud-translate